Apply feedback from the review. Here are the notes from the review:
add video events to ToC

Change all non-code "TrueX" variations to "true[X]"

add an example block of code for actually including TAR in your project -- Roku integration docs has an example of this

mid-roll and pre-roll with the dashes

reintroduce "AD_FETCH_COMPLETED" to the docs -- look for existing example usages in the tvOS docs
- also change Roku event name to "adFetchCompleted"
- if you want to implement your own timeout, do like this: call `init`, then if your desired timeout length has passed before getting `AD_FETCH_COMPLETED`, then call `stop` on the renderer and remove all references to the renderer. Then you may proceed with your stream with regular video ads.

"Another possible event emitted on exit by the TruexAdRenderer is OPT_OUT."
"An event you can optionally listen for is"?

overload the `init` method with one that doesn't have the "creativeURL" as a parameter, and only document that one

make it clear what the terminal events are. and make sure OPT_OUT doesn't sound like a terminal event.
- AD_COMPLETED, USER_CANCEL_STREAM, NO_ADS_AVAILABLE, AD_ERROR

the method descriptions should explain to a user what events to pay attention to after calling the method.

Make sure the `start` method description is clear about the previous bullet ^

replace "US Air Force - Special OPS (TA M) - Q1 2017" with something generic

try doing "```java" at the top of the blocks to properly highlight syntax etc.

change "15 second choice card countdown runs out" to "Choice card countdown runs out"

remove "(or taps I'm done)"

remove " -- set to 0 if the user did not earn an ad free credit or if the user was shown a "skip card"." for AD_COMPLETED

change "should not take any further action until it receives an AD_COMPLETED or AD_ERROR event." to "should not take any further action until it receives a terminal event."

"All following events are used mostly for tracking purposes -- no action is generally required" -- move this to below POPUP_WEBSITE, and make this a real header for another section of events. then remove "(optional)" from all the following events.

change " in a WebView with a close button. Once the user closes this view" to ". Once the user closes the popup"

add descriptions for the parameters of each event, like how AD_STARTED has that

explain that the behavior of TAR will change if USER_CANCEL_STREAM is listened for. Put this in the description for USER_CANCEL_STREAM.

move the VIDEO_[foo] events to a different section to make it clear that these are NOT like ad_started, etc. -- these are really just for any videos that may happen to be in the ad creative.